% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{barcode_columns}
\alias{barcode_columns}
\title{Default set of columns used to construct barcodes}
\format{
An object of class \code{character} of length 7.
}
\usage{
barcode_columns
}
\description{
Shared helper constant listing the Excel columns that are concatenated to
form the unique barcode identifiers across the package.
}
\keyword{internal}
