% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPCR2}
\alias{getPCR2}
\title{Further Process PCR Data and Annotate Plate Matrices}
\usage{
getPCR2(pcr_data)
}
\arguments{
\item{pcr_data}{A list of data frames representing plate layouts from `getPCR`.}
}
\value{
A list containing the annotated plate data frames and a vector of gene names from the 96th position.
}
\description{
Takes the output from `getPCR` function, performs additional processing,
extracts gene names from the 96th position, and annotates the plate matrices with "C-".
}
\examples{
\dontrun{
# Example assumes `pcr_data` is available from `getPCR`.
getPCR2(pcr_data)
}
}
