% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPCR}
\alias{getPCR}
\title{Create Plate Layout for PCR from User-Provided Excel File}
\usage{
getPCR(file_path = NULL)
}
\arguments{
\item{file_path}{A string specifying the path to the Excel file containing PCR data.
The file should be in a format readable by `readxl::read_excel()`.}
}
\value{
A list of data frames, each representing the layout of a PCR plate.
        Rows are labeled A-H and columns 1-12, with each cell containing a unique barcode.
}
\description{
This function processes PCR data from a user-provided Excel file and organizes the data into
plate layouts. It assumes each plate has 96 wells and arranges unique barcode sequences
into an 8x12 matrix corresponding to the layout of a standard PCR plate.
}
\details{
The function reads the user-provided Excel file and selects necessary columns related to the barcode.
It then creates a unified barcode by concatenating these columns with an underscore separator.
After deduplicating barcodes, it calculates the number of plates needed and arranges barcodes into
an 8x12 matrix for each plate. The final list of matrices can be used to guide the setup of PCR plates
in a laboratory setting.
}
\examples{
\dontrun{
file_path <- system.file("extdata", "ResultWithComplementSeq_final_file.xlsx", package = "smFishPlateDesigner")
plates <- getPCR(file_path)
}
}
