% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{processDosageTIV}
\alias{processDosageTIV}
\title{Process Dosage TIV Data from User-Provided Excel File}
\usage{
processDosageTIV(file_path = NULL)
}
\arguments{
\item{file_path}{A string specifying the path to the Excel file to be processed.
The file should be in a format readable by `readxl::read_excel()`.}
}
\value{
A data frame containing the processed dosage data.
}
\description{
This function processes TIV dosage data from an Excel file specified by the user.
The user must provide the path to their Excel file containing the dosage data.
}
\examples{
\dontrun{
file_path <- system.file("extdata", "ResultWithComplementSeq_final_file.xlsx", package = "smFishPlateDesigner")
processed_data <- processDosageTIV(file_path)
}
}
