% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{processFishDataWithoutPrimers}
\alias{processFishDataWithoutPrimers}
\title{Process Fish Data into Plate Layouts from User-Provided Excel File}
\usage{
processFishDataWithoutPrimers(file_path = NULL)
}
\arguments{
\item{file_path}{A string specifying the path to the Excel file containing fish data.
The file should be in a format readable by `readxl::read_excel()`.}
}
\value{
A list containing the plate layouts with annotations. Each layout will correspond
        to the structure of a standard PCR plate, with specific barcode allocations.
}
\description{
This function processes fish data from a user-provided Excel file, creating unique barcodes for each sample
and arranging them into plate layouts with specific annotations. The function is tailored to handle data
where primer information is not required or is pre-processed.
}
\details{
The function reads the specified Excel file, extracting necessary information for barcode creation
and plate layout organization, excluding primer information. It processes this information to output
a list of plate layouts, each corresponding to a unique set of samples and annotations.
}
\examples{
\dontrun{
file_path <- system.file("extdata", "ResultWithComplementSeq_final_file.xlsx", package = "smFishPlateDesigner")
processFishDataWithoutPrimers(file_path)
}
}
